+++ cert/cert.go
@@ -68,10 +68,7 @@ func (config *Cert) requestCertificate() (apiResponse, error) {
     }
 
 	err = json.Unmarshal(body, &r)
-	if err != nil {
-		return r, err
-	}
-	return r, nil
+	return r, err
 }
 
 // WriteCert the certificate to disk
@@ -81,6 +78,9 @@ func (config *Cert) WriteCert(c SignedCertificate) error {
 	if config.OutputFormat == "pem" {
 		pem := []byte(c.Certificate + "\n" + c.PrivateKey + "\n" + c.IssuingCa)
 		pemFile := config.getDir() + config.Hostname + ".pem"
+        if file.Exists(pemFile) && config.OverWrite == false {
+            return errors.New(pemFile + " already exists.")
+        }
         if err := file.WriteFile(pemFile, pem, 0600, config.OverWrite); err != nil {
             return err
         }
@@ -90,15 +90,19 @@ func (config *Cert) WriteCert(c SignedCertificate) error {
 	} else if config.OutputFormat == "cert" {
 		crt := []byte(c.Certificate + "\n" + c.IssuingCa)
 		crtFile := config.getDir() + config.Hostname + ".crt"
-        err := file.WriteFile(crtFile, crt, 0600, config.OverWrite)
-		if err != nil {
+        if file.Exists(crtFile) && config.OverWrite == false {
+            return errors.New(crtFile + " already exists.")
+        }
+        if err := file.WriteFile(crtFile, crt, 0600, config.OverWrite); err != nil {
 			return err
 		}
 
 		key := []byte(c.PrivateKey)
 		keyFile := config.getDir() + config.Hostname + ".key"
-        err = file.WriteFile(keyFile, key, 0600, config.OverWrite)
-		if err != nil {
+        if file.Exists(keyFile) && config.OverWrite == false {
+            return errors.New(keyFile + " already exists.")
+        }
+        if err := file.WriteFile(keyFile, key, 0600, config.OverWrite); err != nil {
 			return err
 		}
 
@@ -124,8 +128,10 @@ func (config *Cert) WriteCert(c SignedCertificate) error {
 		}
 
 		p12File := config.getDir() + config.Hostname + ".p12"
-        err = file.WriteFile(p12File, p12, 0600, config.OverWrite)
-		if err != nil {
+        if file.Exists(p12File) && config.OverWrite == false {
+            return errors.New(p12File + " already exists.")
+        }
+        if err := file.WriteFile(p12File, p12, 0600, config.OverWrite); err != nil {
 			return err
 		}
 	}
